infra: split the widget helpers out of utils.h
utils.h is included by TBuffer.h for qsl(), so every core-model file that
included it pulled QApplication and QWidget in behind it. The dialog
positioning and child-signal helpers move to WidgetUtils.h; utils.h keeps
the core-safe content and no longer includes anything from Qt Widgets.
